bit manipulation

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -11,10 +11,11 @@
 */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int mask;
+	unsigned long int mask;
 	int bit_value;
 
-	if (index >= sizeof(unsigned long int) * 8) {
+	if (index >= sizeof(unsigned long int) * 8)
+	{
         return -1;
     }
