+++ 0x14-bit_manipulation/3-set_bit.c
@@ -11,9 +11,10 @@
 */
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned int mask;
+	unsigned long int mask;
 
-	if (index >= sizeof(unsigned long int) * 8) {
+	if (index >= sizeof(unsigned long int) * 8)
+	{
         return -1;
     }
 
@@ -22,5 +23,4 @@ int set_bit(unsigned long int *n, unsigned int index)
 	*n = *n | mask;
 
     return (1);
-
 }
